Modify: Split `PagesCarousel` from lesson.dart
